Use the config-file format for per-file configuration overrides
